Homework 4/services/scrapers/issuers/requirements.txt
setuptools
aiohttp
aiolimiter
asyncpg
beautifulsoup4
lxml
python-dotenv